备份

+++ src/planner/traj_opt/trajectory_opt.hpp
@@ -19,7 +19,7 @@ public:
         const std::vector<Eigen::Vector2d>& path,
         const RoboState& now,
         bool use_opt,
-        std::optional<std::pair<double, double>> some_no_opt = std::nullopt
+        std::optional<std::pair<int, int>> some_no_opt = std::nullopt
     );
     struct Impl;
     std::unique_ptr<Impl> _impl;
